Add CI capabilities to the repo (#13)
* Add file for experiments in task 1

* Add parameters for task 1 evaluation

* Add script to run task 2

* Add great circle distance to all tests

* Add github actions and fix dependency

* Remove Python 3.9

* Add mock dependency

* Add poetry to better handle dependencies in multiple Python versions

* Remove unexistent Python 3.10

* Skip test requiring raw data

* Move toy instance to conftest for simplicity

Create a folder with toy instances just for the tests

* Fix wrong file in tests

* Delete temporary and test files

* Replace the deprecated np.int with np.int32

* Add FIXME in test that does not work in Github actions

* Add quick Poetry description in quickstart.md

* Add author names to the project

* Fix e-mail mistake in authors

+++ loggibud/v1/baselines/shared/ortools.py
@@ -68,7 +68,7 @@ def solve(
     model = pywrapcp.RoutingModel(manager)
 
     # Unwrap the size index for every point.
-    sizes = np.array([0] + [d.size for d in instance.deliveries], dtype=np.int)
+    sizes = np.array([0] + [d.size for d in instance.deliveries], dtype=np.int32)
 
     def capacity_callback(src):
         src = manager.IndexToNode(src)
@@ -84,7 +84,7 @@ def solve(
 
     # Compute the distance matrix between points.
     logger.info("Computing distance matrix.")
-    distance_matrix = (calculate_distance_matrix_m(locations) * 10).astype(np.int)
+    distance_matrix = (calculate_distance_matrix_m(locations) * 10).astype(np.int32)
 
     def distance_callback(src, dst):
         x = manager.IndexToNode(src)
